src/requirements.txt
pcf8575==0.3
RPi.GPIO==0.7.0
python-statemachine==0.8.0
flask==2.0.2
python-chess==1.999
gevent==21.8.0
python-dotenv==0.19.2